backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==44.0.0
defusedxml==0.7.1
Django==4.2.11
django-bootstrap-icons==0.8.7
django-bootstrap-v5==1.0.11
django-cors-headers==4.6.0
django-cors-middleware==1.5.0
django-environ==0.11.2
django-jquery==3.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.7
jwt==1.3.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==68.2.2
soupsieve==2.5
sqlparse==0.5.0
urllib3==2.2.2
wheel==0.41.2
